make simple decomposition
